sm8350-common: props: Add support for low power adaptive control feature High power consumption issues were reported for some HS/CK/Speaker in close range(1 meter) on hastings platform as they may not have perfect RSSI golden range logic for power control and always request SOC to increase TX power to max power level. Newer hastings SOC support LPA power control to avoid unnecessary higher power consumption.
The change enable this soc feature according to value of property.
And disable it for special remote devices which have been added in BL
when new ACL link is created.

If feature is enabled, controllable power level for remote is L0-L9,
BT controller increase it to L10 or L11 according to RSSI and NAK rate.
Otherwise controllable power level for remote is from 0 to max power
level signified in NVM, most likely L11 for hastings.

CRs-Fixed: 2597191
Change-Id: I354bb34a84aa7a55537e3a595daac54c46d90603